fix : 일정 페이지 이상에서 데이터를 읽어올 수 없는 버그 해결 #36

+++ OpenMarket/OpenMarket/Extension/UInt+extension.swift
@@ -7,7 +7,7 @@
 
 import Foundation
 
-extension UInt {
+extension Double {
     func priceFormat(currency : String?) -> String? {
         let numberFormatter = NumberFormatter()
         numberFormatter.numberStyle = .decimal
